+++ packages/google_sign_in/google_sign_in_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 7.2.3
+
+* Removed mockito-inline and updated to mockito-core 5.17.0.
+
 ## 7.2.2
 
 * Updates Java compatibility version to 17 and minimum supported SDK version to Flutter 3.35/Dart 3.9.